requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-openai
langchain-core
langchain-community
bs4
tabulate